new-backend/requirements.txt
flask==3.1.1
flask-cors==3.0.10
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

# Supabase Integration
supabase==2.3.4
python-jose[cryptography]==3.3.0

# OpenAI Integration
openai==1.3.5

# Core AI Dependencies
numpy==1.24.3
pillow==10.1.0
opencv-python-headless==4.8.1.78

# Advanced AI Dependencies
faiss-cpu==1.7.4
timm==0.9.12
transformers==4.35.0
torch==2.1.0

# SCIN Dataset Integration
gcsfs==2023.10.0
google-auth==2.23.4

# Custom Models and Real-time Analysis
scikit-learn==1.3.2
joblib==1.3.2

# Performance and Monitoring
psutil==5.9.6 